requirements.txt
# Core Twitter API
tweepy==4.14.0

# Web framework for status tracker
fastapi==0.104.1
uvicorn==0.24.0

# Scheduling
apscheduler==3.10.4

# Trend extraction
pytrends==4.9.2
langdetect==1.0.9

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Database
aiosqlite==0.19.0

# File operations
aiofiles==23.2.0

# Timezone support
pytz==2023.3

# CLI
click==8.1.7

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging
python-json-logger==2.0.7

# Machine Learning for relevance scoring
scikit-learn==1.3.2
numpy==1.26.2

# Text similarity and deduplication
python-Levenshtein==0.23.0

# System monitoring
psutil==5.9.6

# UUID generation (built-in, no install needed)
# uuid

# JSON handling (built-in, no install needed)
# json

# Path operations (built-in, no install needed)
# pathlib
